bundle all packages
