+++ .eslintignore
@@ -8,6 +8,7 @@ prism.ts
 charm.ts
 pnpm-lock.yaml
 generated-dmmf.ts
+*.d.ts
 
 # client
 packages/client/generator-build/
