+++ packages/cli/src/Studio.ts
@@ -1,7 +1,16 @@
 import Debug from '@prisma/debug'
 import { enginesVersion } from '@prisma/engines'
-import { arg, Command, format, getConfig, getDirectUrl, HelpError, isError, loadEnvFile } from '@prisma/internals'
-import { resolveUrl } from '@prisma/internals/dist/engine-commands/getConfig'
+import {
+  arg,
+  Command,
+  format,
+  getConfig,
+  getDirectUrl,
+  HelpError,
+  isError,
+  loadEnvFile,
+  resolveUrl,
+} from '@prisma/internals'
 import { getSchemaPathAndPrint } from '@prisma/migrate'
 import { StudioServer } from '@prisma/studio-server'
 import fs from 'fs'
