+++ packages/debug/package.json
@@ -19,7 +19,10 @@
     "esbuild": "0.19.5",
     "jest": "29.7.0",
     "jest-junit": "16.0.0",
-    "typescript": "5.2.2"
+    "@types/debug": "4.1.10",
+    "typescript": "5.2.2",
+    "debug": "4.3.4",
+    "strip-ansi": "6.0.1"
   },
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
@@ -31,10 +34,6 @@
     "README.md",
     "dist"
   ],
-  "dependencies": {
-    "@types/debug": "4.1.10",
-    "debug": "4.3.4",
-    "strip-ansi": "6.0.1"
-  },
+  "dependencies": {},
   "sideEffects": false
 }
