+++ packages/engines/tsconfig.build.json
@@ -1,7 +1,7 @@
 {
   "extends": "../../tsconfig.build.bundle.json",
   "compilerOptions": {
-    "outDir": "dist"
+    "outDir": "declaration"
   },
   "include": ["src"]
 }
