+++ packages/fetch-engine/package.json
@@ -25,9 +25,7 @@
     "del": "6.1.1",
     "jest": "29.7.0",
     "strip-ansi": "6.0.1",
-    "typescript": "5.2.2"
-  },
-  "dependencies": {
+    "typescript": "5.2.2",
     "@prisma/debug": "workspace:*",
     "@prisma/get-platform": "workspace:*",
     "execa": "5.1.1",
@@ -46,6 +44,7 @@
     "temp-dir": "2.0.0",
     "tempy": "1.0.1"
   },
+  "dependencies": {},
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
     "build": "node -r esbuild-register helpers/build.ts",
