+++ packages/generator-helper/package.json
@@ -14,12 +14,7 @@
   "author": "Tim Suchanek <suchanek@prisma.io>",
   "homepage": "https://www.prisma.io",
   "bugs": "https://github.com/prisma/prisma/issues",
-  "dependencies": {
-    "@prisma/debug": "workspace:*",
-    "@types/cross-spawn": "6.0.4",
-    "cross-spawn": "7.0.3",
-    "kleur": "4.1.5"
-  },
+  "dependencies": {},
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
     "build": "node -r esbuild-register helpers/build.ts",
@@ -27,6 +22,10 @@
     "prepublishOnly": "pnpm run build"
   },
   "devDependencies": {
+    "@prisma/debug": "workspace:*",
+    "@types/cross-spawn": "6.0.4",
+    "cross-spawn": "7.0.3",
+    "kleur": "4.1.5",
     "@swc-node/register": "1.6.8",
     "@swc/core": "1.3.95",
     "@swc/jest": "0.2.29",
