+++ packages/get-platform/helpers/build.ts
@@ -1,3 +1,13 @@
 import { build } from '../../../helpers/compile/build'
 
-void build([{ name: 'default' }])
+void build([
+  {
+    name: 'default',
+    entryPoints: ['src/index.ts'],
+    outfile: 'dist/index',
+    bundle: true,
+    minify: true,
+    sourcemap: true,
+    emitTypes: true,
+  },
+])
