+++ packages/get-platform/package.json
@@ -22,9 +22,7 @@
     "benchmark": "2.1.4",
     "jest": "29.7.0",
     "jest-junit": "16.0.0",
-    "typescript": "5.2.2"
-  },
-  "dependencies": {
+    "typescript": "5.2.2",
     "@prisma/debug": "workspace:*",
     "escape-string-regexp": "4.0.0",
     "execa": "5.1.1",
@@ -36,6 +34,7 @@
     "terminal-link": "2.1.1",
     "ts-pattern": "5.0.5"
   },
+  "dependencies": {},
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
     "build": "node -r esbuild-register helpers/build.ts",
