+++ packages/internals/src/index.ts
@@ -28,6 +28,7 @@ export { arg, format, isError } from './cli/utils'
 export { ClientEngineType, DEFAULT_CLIENT_ENGINE_TYPE, getClientEngineType } from './client/getClientEngineType'
 export { credentialsToUri, protocolToConnectorType, uriToCredentials } from './convertCredentials'
 export * from './engine-commands'
+export { resolveUrl } from './engine-commands/getConfig'
 export { Generator } from './Generator'
 export type { GeneratorPaths } from './get-generators/generatorResolvers/generatorResolvers'
 export type { ProviderAliases } from './get-generators/getGenerators'
@@ -87,6 +88,7 @@ export { serializeQueryEngineName } from './utils/serializeQueryEngineName'
 export { setClassName } from './utils/setClassName'
 export type { Position } from './utils/trimBlocksFromSchema'
 export { trimBlocksFromSchema, trimNewLine } from './utils/trimBlocksFromSchema'
+export type { LoadedEnv } from './utils/tryLoadEnvs'
 export { tryLoadEnvs } from './utils/tryLoadEnvs'
 export { vercelPkgPathRegex } from './utils/vercelPkgPathRegex'
 export { warnOnce } from './warnOnce'
