+++ packages/migrate/package.json
@@ -33,19 +33,11 @@
     "jest-junit": "16.0.0",
     "mock-stdin": "1.0.0",
     "tempy": "1.0.1",
-    "typescript": "5.2.2"
-  },
-  "peerDependencies": {
-    "@prisma/generator-helper": "*",
-    "@prisma/internals": "*"
-  },
-  "dependencies": {
+    "typescript": "5.2.2",
     "@prisma/debug": "workspace:*",
     "@prisma/get-platform": "workspace:*",
     "@sindresorhus/slugify": "1.1.2",
-    "arg": "5.0.2",
     "execa": "5.1.1",
-    "fp-ts": "2.16.1",
     "get-stdin": "8.0.0",
     "has-yarn": "2.1.0",
     "indent-string": "4.0.0",
@@ -57,10 +49,18 @@
     "ora": "5.4.1",
     "pg": "8.11.3",
     "pkg-up": "3.1.0",
-    "prompts": "2.4.2",
     "strip-ansi": "6.0.1",
     "strip-indent": "3.0.0",
     "ts-pattern": "5.0.5"
+  },
+  "peerDependencies": {
+    "@prisma/generator-helper": "*",
+    "@prisma/internals": "*"
+  },
+  "dependencies": {
+    "arg": "5.0.2",
+    "fp-ts": "2.16.1",
+    "prompts": "2.4.2"
   },
   "scripts": {
     "dev": "DEV=true node -r esbuild-register helpers/build.ts",
